Fix error handling in hazmat/io.py
On an error, it was trying to call godot_print_error, but with the
wrong signature. So no python errors would actually get output.

As the note in the source says, godot_print_error takes 4 args: descr,
func, file, line. So GodotIO.write/flush would need to call it like
that, but we don't have func/file/line here. And also, python calls
write() at fairly random points with substrings of the actual
message, so trying to structure the output with
godot_print_error (which prints "ERROR: <func>: <msg>\n  At:
<file>:<line>") doesn't work if you just have substrings of the thing
getting printed.

So I just switched it to use lib.godot_print which works OK.